requirements.txt
django == 4.2
psycopg2-binary == 2.9.6
psycopg2 == 2.8.6
boto3 == 1.26.146
djangorestframework == 3.14.0
neo4j == 5.11.0